Removing NFTpage
